# talentwood-backend-develop/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Allow the user to login",
                "description": "Returns the validated user data along with the api token",
                "operationId": "1726f27d58c023ccef2df969fa897475",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Encoded User Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user/loginwithotp": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Allow the user to login with otp",
                "description": "Returns the validated user data along with the api token",
                "operationId": "47c4d1c457b89d1c29442d32a024a42a",
                "parameters": [
                    {
                        "name": "userphone",
                        "in": "query",
                        "description": "User mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userOtp",
                        "in": "query",
                        "description": "OTP",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user/attributes": {
            "get": {
                "tags": [
                    "userAttributes"
                ],
                "summary": "Get all the user attributes details",
                "description": "Returns tall the user attributes details",
                "operationId": "fa088d969105884349d00e5087f13b08",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}